# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aegra",
    "description": "Production-ready Agent Protocol server",
    "version": "0.17.0"
  },
  "paths": {
...
    "/assistants": {
      "get": {
        "tags": [
          "Assistants"
        ],
        "summary": "List Assistants",
        "description": "List all assistants owned by the authenticated user.\n\nReturns every assistant without filtering. Use the search endpoint for\nfiltered queries.",
        "operationId": "list_assistants_assistants_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Assistants"
        ],
        "summary": "Create Assistant",
        "description": "Create a new assistant.\n\nAn assistant is a configured instance of a graph. Provide a `graph_id`\nreferencing a graph defined in your `aegra.json`. If `assistant_id` is\nomitted, one is auto-generated. Set `if_exists` to `\"do_nothing\"` for\nidempotent creation.",
        "operationId": "create_assistant_assistants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assistant"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assistants/search": {
      "post": {
        "tags": [
          "Assistants"
        ],
        "summary": "Search Assistants",
        "description": "Search assistants with filters.\n\nFilter by name, description, graph ID, or metadata. Results are paginated\nvia `limit` and `offset`; a full page sets the `X-Pagination-Next` cursor\nheader. Pass `select` to return only the listed fields.\n\nDeclared without a response model because `select` makes the row shape\ndynamic \u2014 full entities when omitted, projected dicts when given.",
        "operationId": "search_assistants_assistants_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assistants/count": {
      "post": {
        "tags": [
          "Assistants"
        ],
        "summary": "Count Assistants",
        "description": "Count assistants matching the given filters.\n\nAccepts the same filter parameters as the search endpoint but returns only\nthe total count.",
        "operationId": "count_assistants_assistants_count_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Response Count Assistants Assistants Count Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assistants/{assistant_id}": {
      "get": {
        "tags": [
          "Assistants"
        ],
        "summary": "Get Assistant",
        "description": "Get an assistant by its ID.\n\nReturns the latest version of the assistant. Returns 404 if the assistant\ndoes not exist or does not belong to the authenticated user.",
        "operationId": "get_assistant_assistants__assistant_id__get",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assistant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assistant"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Assistants"
        ],
        "summary": "Update Assistant",
        "description": "Update an assistant by its ID.\n\nPartial update: only fields included in the request body are changed.\nCreates a new version of the assistant.",
        "operationId": "update_assistant_assistants__assistant_id__patch",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assistant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assistant"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Assistants"
        ],
        "summary": "Delete Assistant",
        "description": "Delete an assistant by its ID.\n\nPermanently removes the assistant and all of its versions. Set\n`delete_threads` to also remove every thread created against it. This\naction cannot be undone.",
        "operationId": "delete_assistant_assistants__assistant_id__delete",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assistant Id"
            }
          },
          {
            "name": "delete_threads",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also delete every thread created against this assistant.",
              "default": false,
              "title": "Delete Threads"
            },
            "description": "Also delete every thread created against this assistant."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assistants/{assistant_id}/latest": {
      "post": {
        "tags": [
          "Assistants"
        ],
        "summary": "Set Assistant Latest",
        "description": "Pin a specific version as the latest version of an assistant.\n\nAfter calling this endpoint, the assistant will use the specified version's\nconfiguration when executing runs.",
        "operationId": "set_assistant_latest_assistants__assistant_id__latest_post",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assistant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_set_assistant_latest_assistants__assistant_id__latest_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assistant"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assistants/{assistant_id}/versions": {
      "post": {
        "tags": [
          "Assistants"
        ],
        "summary": "List Assistant Versions",
        "description": "List all versions of an assistant.\n\nReturns versions ordered from newest to oldest, paginated via `limit` and\n`offset`. Each version captures the assistant's configuration at the time\nof creation or update.",
        "operationId": "list_assistant_versions_assistants__assistant_id__versions_post",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assistant Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AssistantVersionsRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Assistant"
                  },
                  "title": "Response List Assistant Versions Assistants  Assistant Id  Versions Post"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assistants/{assistant_id}/schemas": {
      "get": {
        "tags": [
          "Assistants"
        ],
        "summary": "Get Assistant Schemas",
        "description": "Get the JSON schemas for an assistant's graph.\n\nReturns the input, output, state, and config schemas derived from the\nunderlying graph's type annotations.",
        "operationId": "get_assistant_schemas_assistants__assistant_id__schemas_get",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assistant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSchemas"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assistants/{assistant_id}/graph": {
      "get": {
        "tags": [
          "Assistants"
        ],
        "summary": "Get Assistant Graph",
        "description": "Get the graph structure for visualization.\n\nReturns a JSON representation of the graph's nodes and edges suitable for\nrendering in graph visualizers. Use `xray` to expand subgraph nodes into\ntheir internal structure.",
        "operationId": "get_assistant_graph_assistants__assistant_id__graph_get",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assistant Id"
            }
          },
          {
            "name": "xray",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Expand subgraph nodes. Pass true or a depth integer to control nesting.",
              "title": "Xray"
            },
            "description": "Expand subgraph nodes. Pass true or a depth integer to control nesting."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assistants/{assistant_id}/subgraphs": {
      "get": {
        "tags": [
          "Assistants"
        ],
        "summary": "Get Assistant Subgraphs",
        "description": "Get subgraphs of an assistant.\n\nReturns the subgraph definitions used by this assistant's graph. Set\n`recurse=true` to include deeply nested subgraphs, or filter to a single\nnamespace.",
        "operationId": "get_assistant_subgraphs_assistants__assistant_id__subgraphs_get",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assistant Id"
            }
          },
          {
            "name": "recurse",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Recursively include nested subgraphs.",
              "default": false,
              "title": "Recurse"
            },
            "description": "Recursively include nested subgraphs."
          },
          {
            "name": "namespace",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to a specific subgraph namespace.",
              "title": "Namespace"
            },
            "description": "Filter to a specific subgraph namespace."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assistants/{assistant_id}/subgraphs/{namespace}": {
      "get": {
        "tags": [
          "Assistants"
        ],
        "summary": "Get Assistant Subgraphs By Namespace",
        "description": "Get subgraphs of an assistant under a specific namespace.\n\nPath form of the namespace filter \u2014 the SDK uses this variant whenever a\nnamespace is supplied, and the query form when it is not.",
        "operationId": "get_assistant_subgraphs_by_namespace_assistants__assistant_id__subgraphs__namespace__get",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assistant Id"
            }
          },
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            }
          },
          {
            "name": "recurse",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Recursively include nested subgraphs.",
              "default": false,
              "title": "Recurse"
            },
            "description": "Recursively include nested subgraphs."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssistantCreate": {
        "properties": {
          "assistant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assistant Id",
            "description": "Unique assistant identifier (auto-generated if not provided)"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Human-readable assistant name (auto-generated if not provided)"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Assistant description"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config",
            "description": "Assistant configuration"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "Assistant context"
          },
          "graph_id": {
            "type": "string",
            "title": "Graph Id",
            "description": "LangGraph graph ID from aegra.json"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Metadata to use for searching and filtering assistants."
          },
          "if_exists": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "raise",
                  "do_nothing"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "If Exists",
            "description": "On conflict: 'raise' returns 409, 'do_nothing' returns the existing assistant.",
            "default": "raise"
          }
        },
        "type": "object",
        "required": [
          "graph_id"
        ],
        "title": "AssistantCreate",
        "description": "Request model for creating assistants"
      },
      "AssistantList": {
        "properties": {
          "assistants": {
            "items": {
              "$ref": "#/components/schemas/Assistant"
            },
            "type": "array",
            "title": "Assistants"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "assistants",
          "total"
        ],
        "title": "AssistantList",
        "description": "Response model for listing assistants"
      },
...
      "Body_set_assistant_latest_assistants__assistant_id__latest_post": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version",
            "description": "The version number to set as latest"
          }
        },
        "type": "object",
        "required": [
          "version"
        ],
        "title": "Body_set_assistant_latest_assistants__assistant_id__latest_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  },
...
}